+++ client/apps/game/src/ui/layouts/world.tsx
@@ -3,8 +3,8 @@ import { useUIStore } from "@/hooks/store/use-ui-store";
 import { LoadingStateKey } from "@/hooks/store/use-world-loading";
 import { LoadingOroborus } from "@/ui/modules/loading-oroborus";
 import { LoadingScreen } from "@/ui/modules/loading-screen";
-import { getAllStructuresFromToriiClient } from "@bibliothecadao/eternum";
 import { useDojo, usePlayerStructures } from "@bibliothecadao/react";
+import { getAllStructuresFromToriiClient } from "@bibliothecadao/torii-client";
 import { Leva } from "leva";
 import { lazy, Suspense, useCallback, useEffect, useRef, useState } from "react";
 import { env } from "../../../env";
