+++ packages/core/src/utils/structure.ts
@@ -9,12 +9,10 @@ import {
   TickIds,
 } from "@bibliothecadao/types";
 import { ComponentValue, Entity, getComponentValue } from "@dojoengine/recs";
-import { Clause, PatternMatching, Query, ToriiClient } from "@dojoengine/torii-wasm";
 import { getEntityIdFromKeys } from "@dojoengine/utils";
 import { shortString } from "starknet";
 import { configManager } from "../managers";
 import { getEntityName } from "./entities";
-import { getResourcesFromToriiEntity } from "./resources";
 import { currentTickCount } from "./utils";
 
 export const getStructureAtPosition = (
@@ -110,202 +108,4 @@ export const getStructureTypeName = (structureType: StructureType) => {
     default:
       return "Unknown";
   }
-};
-
-export const getStructureFromToriiClient = async (toriiClient: ToriiClient, entityId: ID) => {
-  const query: Query = {
-    limit: 1,
-    offset: 0,
-    clause: {
-      Keys: {
-        keys: [entityId.toString()],
-        pattern_matching: "FixedLen" as PatternMatching,
-        models: [],
-      },
-    },
-    dont_include_hashed_keys: false,
-    order_by: [],
-    entity_models: ["s1_eternum-Structure", "s1_eternum-Resource"],
-    entity_updated_after: 0,
-  };
-
-  const entities = await toriiClient.getEntities(query, false);
-  const entity = Object.keys(entities)[0] as Entity;
-
-  if (!entity) return;
-
-  return {
-    structure: getStructureFromToriiEntity(entities[entity]["s1_eternum-Structure"]),
-    resources: getResourcesFromToriiEntity(entities[entity]["s1_eternum-Resource"]),
-  };
-};
-
-export const getAllStructuresFromToriiClient = async (toriiClient: ToriiClient, ownedBy?: string) => {
-  const clause: Clause = !ownedBy
-    ? {
-        Keys: {
-          keys: [undefined], // matches any key
-          pattern_matching: "FixedLen" as PatternMatching,
-          models: ["s1_eternum-Structure"], // specify the model you want to query
-        },
-      }
-    : {
-        Member: {
-          model: "s1_eternum-Structure",
-          member: "owner",
-          operator: "Eq",
-          value: { Primitive: { ContractAddress: ownedBy } },
-        },
-      };
-
-  const query: Query = {
-    limit: 1000,
-    offset: 0,
-    clause,
-    dont_include_hashed_keys: false,
-    order_by: [],
-    entity_models: ["s1_eternum-Structure"],
-    entity_updated_after: 0,
-  };
-
-  const entities = await toriiClient.getEntities(query, false);
-  const result = Object.keys(entities).map((entity) => {
-    const structure = getStructureFromToriiEntity(entities[entity]["s1_eternum-Structure"]);
-    return {
-      entityId: structure.entity_id,
-      owner: structure.owner,
-      position: { col: structure.base.coord_x, row: structure.base.coord_y },
-    };
-  });
-  return result;
-};
-
-export const getFirstStructureFromToriiClient = async (toriiClient: ToriiClient, ownedBy?: string) => {
-  const clause: Clause = !ownedBy
-    ? {
-        Keys: {
-          keys: [undefined], // matches any key
-          pattern_matching: "FixedLen" as PatternMatching,
-          models: ["s1_eternum-Structure"], // specify the model you want to query
-        },
-      }
-    : {
-        Composite: {
-          operator: "And",
-          clauses: [
-            {
-              Keys: {
-                keys: [undefined], // matches any key
-                pattern_matching: "FixedLen" as PatternMatching,
-                models: ["s1_eternum-Structure"], // specify the model you want to query
-              },
-            },
-            {
-              Member: {
-                model: "s1_eternum-Structure",
-                member: "owner",
-                operator: "Eq",
-                value: { Primitive: { ContractAddress: ownedBy } },
-              },
-            },
-          ],
-        },
-      };
-
-  const query: Query = {
-    limit: 1,
-    offset: 0,
-    clause,
-    dont_include_hashed_keys: false,
-    order_by: [],
-    entity_models: ["s1_eternum-Structure"],
-    entity_updated_after: 0,
-  };
-
-  const entities = await toriiClient.getEntities(query, false);
-  const realmEntity = Object.keys(entities)[0] as Entity;
-
-  if (!realmEntity) {
-    return;
-  }
-
-  const structure = getStructureFromToriiEntity(entities[realmEntity]["s1_eternum-Structure"]);
-
-  return {
-    entityId: structure.entity_id,
-    owner: structure.owner,
-    position: { col: structure.base.coord_x, row: structure.base.coord_y },
-  };
-};
-
-export const getStructureFromToriiEntity = (entityData: any) => {
-  const structure: ComponentValue<ClientComponents["Structure"]["schema"]> = {
-    entity_id: entityData.entity_id?.value,
-    owner: ContractAddress(entityData.owner?.value),
-    category: entityData.category?.value,
-    resources_packed: entityData.resources_packed?.value,
-    base: {
-      troop_guard_count: entityData.base?.value?.troop_guard_count?.value,
-      troop_explorer_count: entityData.base?.value?.troop_explorer_count?.value,
-      troop_max_guard_count: entityData.base?.value?.troop_max_guard_count?.value,
-      troop_max_explorer_count: entityData.base?.value?.troop_max_explorer_count?.value,
-      created_at: entityData.base?.value?.created_at?.value,
-      category: entityData.base?.value?.category?.value,
-      coord_x: entityData.base?.value?.coord_x?.value,
-      coord_y: entityData.base?.value?.coord_y?.value,
-      level: entityData.base?.value?.level?.value,
-    },
-    troop_guards: {
-      delta: {
-        category: entityData.troop_guards?.value?.delta?.value?.category?.value?.option,
-        tier: entityData.troop_guards?.value?.delta?.value?.tier?.value?.option,
-        count: entityData.troop_guards?.value?.delta?.value?.count?.value,
-        stamina: {
-          amount: entityData.troop_guards?.value?.delta?.value?.stamina?.value?.amount?.value,
-          updated_tick: entityData.troop_guards?.value?.delta?.value?.stamina?.value?.updated_tick?.value,
-        },
-      },
-      charlie: {
-        category: entityData.troop_guards?.value?.charlie?.value?.category?.value?.option,
-        tier: entityData.troop_guards?.value?.charlie?.value?.tier?.value?.option,
-        count: entityData.troop_guards?.value?.charlie?.value?.count?.value,
-        stamina: {
-          amount: entityData.troop_guards?.value?.charlie?.value?.stamina?.value?.amount?.value,
-          updated_tick: entityData.troop_guards?.value?.charlie?.value?.stamina?.value?.updated_tick?.value,
-        },
-      },
-      bravo: {
-        category: entityData.troop_guards?.value?.bravo?.value?.category?.value?.option,
-        tier: entityData.troop_guards?.value?.bravo?.value?.tier?.value?.option,
-        count: entityData.troop_guards?.value?.bravo?.value?.count?.value,
-        stamina: {
-          amount: entityData.troop_guards?.value?.bravo?.value?.stamina?.value?.amount?.value,
-          updated_tick: entityData.troop_guards?.value?.bravo?.value?.stamina?.value?.updated_tick?.value,
-        },
-      },
-      alpha: {
-        category: entityData.troop_guards?.value?.alpha?.value?.category?.value?.option,
-        tier: entityData.troop_guards?.value?.alpha?.value?.tier?.value?.option,
-        count: entityData.troop_guards?.value?.alpha?.value?.count?.value,
-        stamina: {
-          amount: entityData.troop_guards?.value?.alpha?.value?.stamina?.value?.amount?.value,
-          updated_tick: entityData.troop_guards?.value?.alpha?.value?.stamina?.value?.updated_tick?.value,
-        },
-      },
-      delta_destroyed_tick: entityData.troop_guards?.value?.delta_destroyed_tick?.value,
-      charlie_destroyed_tick: entityData.troop_guards?.value?.charlie_destroyed_tick?.value,
-      bravo_destroyed_tick: entityData.troop_guards?.value?.bravo_destroyed_tick?.value,
-      alpha_destroyed_tick: entityData.troop_guards?.value?.alpha_destroyed_tick?.value,
-    },
-    troop_explorers: entityData.troop_explorers?.value || [],
-    metadata: {
-      realm_id: entityData.metadata?.value?.realm_id?.value,
-      order: entityData.metadata?.value?.order?.value,
-      has_wonder: entityData.metadata?.value?.has_wonder?.value,
-      villages_count: entityData.metadata?.value?.villages_count?.value,
-      village_realm: entityData.metadata?.value?.village_realm?.value,
-    },
-  };
-
-  return structure;
 };
