+++ packages/torii-client/.npmignore
@@ -1,0 +1,7 @@
+*
+
+!dist/**
+!package.json
+!readme.md
+!changelog.md
+!bin
