+++ packages/torii-client/jest.config.js
@@ -1,0 +1,5 @@
+// jest.config.js
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "node",
+};
