+++ packages/torii-client/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@bibliothecadao/torii-client",
+  "version": "0.2.0",
+  "description": "Torii Client Queries for Eternum",
+  "license": "MIT",
+  "author": "Loaf",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
+  "main": "./dist/index.js",
+  "scripts": {
+    "build": "tsup",
+    "gen:docs": "typedoc --plugin typedoc-plugin-markdown",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "test": "vitest"
+  },
+  "dependencies": {
+    "@dojoengine/core": "catalog:",
+    "@dojoengine/sdk": "catalog:",
+    "@bibliothecadao/dojo": "workspace:*",
+    "@bibliothecadao/provider": "workspace:*",
+    "@bibliothecadao/types": "workspace:*"
+  },
+  "devDependencies": {
+    "@types/node": "^20.11.10",
+    "tsup": "^8.0.2",
+    "typedoc": "^0.26.11",
+    "typedoc-plugin-markdown": "^4.2.10",
+    "typescript": "^5.4.4",
+    "vitest": "^2.0.5"
+  }
+}
