+++ packages/torii-client/readme.md
@@ -1,0 +1,3 @@
+# Eternum SDK
+
+The Eternum SDK contains constants and a provider to enable the creation of clients easily.
