+++ packages/torii-client/src/queries/index.ts
@@ -1,0 +1,3 @@
+export * from "./address";
+export * from "./army";
+export * from "./structure";
