+++ packages/torii-client/tsup.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src/index.ts"],
+  target: "esnext",
+  format: ["esm"],
+  dts: true,
+  sourcemap: true,
+  clean: true,
+  minify: true,
+  external: ["@dojoengine/torii-client"],
+});
