+++ packages/torii-client/typedoc.json
@@ -1,0 +1,7 @@
+{
+  "plugin": ["typedoc-plugin-markdown"],
+  "out": "../../../eternum-docs/docs/pages/types",
+  "entryPoints": ["src/index.ts"],
+  "outputFileStrategy": "members",
+  "flattenOutputFiles": true
+}
